client/pyproject.toml
[tool.poetry]
name = "xkcd-explorer"
version = "0.1.0"
description = "XKCD Explorer Python Client"
authors = ["odehaene <olivier.dehaene@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
protobuf = "^3.19.3"
grpcio = "^1.46.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
